fastapi_app/requirements.txt
fastapi
uvicorn
requests
pydantic-settings
google-generativeai
beautifulsoup4
lxml
playwright